fix: Repository Type Integer to Long

+++ src/main/java/mangopot_backend/mangopot_backend/project/repository/JpaProjectInterestRelationRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 
-public interface JpaProjectInterestRelationRepository extends JpaRepository<ProjectInterestRelation, Integer>{
+public interface JpaProjectInterestRelationRepository extends JpaRepository<ProjectInterestRelation, Long>{
 //    List<Project> findByInterest(Interest interest);
     @Query(value = "select p from Project as p join ProjectInterestRelation as pir" +
             " on p.prjId=pir.project.prjId where pir.interest.inId in (:interests)")
